% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_jcr.R
\name{add_jcr}
\alias{add_jcr}
\alias{wos.jcr-class}
\alias{plot.wos.jcr}
\title{Add JCR data to a bibliographic database}
\usage{
add_jcr(db, jcrdb)

\method{plot}{wos.jcr}(
  x,
  filter,
  plot = TRUE,
  all = FALSE,
  ask = plot && all && interactive(),
  ...
)
}
\arguments{
\item{db}{a bibliographic database (a \linkS4class{wos.db} object; typically the
output of the function \code{\link[=db_bib]{db_bib()}}).}

\item{jcrdb}{a JCR database (a \linkS4class{jcr.db} object; typically the
output of the function \code{\link[=db_jcr]{db_jcr()}}).}

\item{x}{a \linkS4class{wos.jcr} object.}

\item{filter}{vector of document identifiers (usually a result of \code{\link[=get_id_docs]{get_id_docs()}}).}

\item{plot}{logical; if \code{TRUE} (default), the plots are drawn, otherwise only
the ggplot2 object is (invisibly) returned.}

\item{all}{logical; if \code{TRUE}, function \code{\link[=plot.wos.db]{plot.wos.db()}} is called (additional
parameters \code{...} are passed to this function), otherwise only a JCR metrics
plot is generated.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot
(see \code{\link{par}(ask=.)}).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An S3 object of \link{class} \code{wos.jcr}.
A \linkS4class{wos.db} object with additional components \code{JCRSour} and \code{JCRCatSour}.
}
\description{
Extends the bibliographic database by adding JCR metrics to sources, per year
and WoS category.
}
\seealso{
\link{db_bib}, \link{db_jcr}.
}
