% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adist_authors.R
\name{adist_authors}
\alias{adist_authors}
\title{Approximate Authors' Names Distances}
\usage{
adist_authors(authors, table, weight = c(family = 0.9, name = 0.1))
}
\arguments{
\item{authors}{author names to be approximately matched.}

\item{table}{lookup table for matching.}

\item{weight}{weights associated with differences in family names and in first names.}
}
\description{
Compute an approximate string distance between character vectors containing authors' names
(of the form \code{"family name, first name"}).
The distance is a weighted proportion of the differences between (capitalized)
family names and between first name initial letters.
}
\seealso{
\code{\link{match_authors}}.
}
