% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos.R
\name{CreateDB}
\alias{CreateDB}
\alias{wos.db-class}
\alias{CreateDB.wos.data}
\alias{print.wos.db}
\alias{plot.wos.db}
\title{Bibliographic data base creation}
\usage{
CreateDB(data, ...)

\method{CreateDB}{wos.data}(data, label = "", progress = interactive(), verbose = FALSE, ...)

\method{print}{wos.db}(x, ...)

\method{plot}{wos.db}(
  x,
  filter,
  which = 1:3,
  plot = TRUE,
  warning = FALSE,
  ask = plot && (length(which) > 1) && interactive(),
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} with bibliographic data.}

\item{...}{further arguments passed to or from other methods.}

\item{label}{character string describing the data.}

\item{progress}{logical; if \code{TRUE} the progress is printed.}

\item{verbose}{logical; if \code{TRUE} additional information is printed.}

\item{x}{an object used to select a method.}

\item{filter}{vector of document identifiers (usually a result of \code{\link{get.idDocs}}).}

\item{which}{if a subset of the plots is required,
specify a subset of the numbers \code{1:3}.}

\item{plot}{logical; if \code{TRUE} (default), the plots are drawn, otherwise only
the list of ggplot2 objects is (invisibly) returned.}

\item{warning}{logical; if \code{FALSE} (default), warnings are ignored.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot
(see \code{\link{par}(ask=.)}).}
}
\value{
An S3 object of \link{class} \code{wos.db}.
A \code{list} with the following components:
\code{Docs}, \code{Authors}, \code{AutDoc}, \code{OI}, \code{OIDoc}, \code{RI}, \code{RIDoc},
\code{Categories}, \code{CatSour}, \code{Areas}, \code{AreaSour}, \code{Addresses}, \code{AddAutDoc},
\code{Affiliations}, \code{AffDoc}, \code{Sources}, \code{WSIndex}, \code{SourWSI}, \code{label} and
\code{export.date}.
}
\description{
\code{CreateDB.wos.data()} converts a \code{data.frame} with WoS variables
(as returned by \code{\link{ImportSources.wos}}; see \code{\link{wosdf}}) into
a relational database (a list of data.frames).
}
\examples{
db <- CreateDB(wosdf)
str(db, 1)
print(db)
summary(db)
}
\seealso{
\code{\link{wosdf}}, \code{\link{ImportSources.wos}}.

\code{\link{summary.wos.db}}.

\code{\link{plot.summary.wos.db}}, \code{\link{plot.summary.year}}
}
