% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos.R
\name{CreateDB.wos}
\alias{CreateDB.wos}
\alias{wos.db-class}
\alias{print.wos.db}
\alias{plot.wos.db}
\title{WoS bibliographic data base}
\usage{
CreateDB.wos(wosdf, label = "", print = interactive())

\method{print}{wos.db}(x, ...)

\method{plot}{wos.db}(x, filter, which = 1:3, plot.summary = FALSE,
  which.summary = 1:5, nmax = 10, plot.summary.year = FALSE,
  which.summary.year = 1:3, accuracy = 0.1, ask = (length(which) +
  plot.summary * length(which.summary) + plot.summary.year *
  length(which.summary.year) > 1) && dev.interactive(), ...)
}
\arguments{
\item{wosdf}{data.frame with WoS data... (as returned by \code{\link{ImportSources.wos}}).}

\item{print}{logical; if \code{TRUE} the progress is printed.}

\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{filter}{vector of document identifiers (usually a result of \code{\link{get.idDocs}}).}

\item{which}{if a subset of the plots is required,
specify a subset of the numbers \code{1:3}.}

\item{plot.summary}{logical; if \code{TRUE}, \code{\link{plot.summary.wos.db}} is called.}

\item{which.summary}{indices of the required plots in \code{\link{plot.summary.wos.db}}.}

\item{nmax}{number of top levels (see \code{\link{summary.wos.db}}).}

\item{plot.summary.year}{logical; if \code{TRUE}, \code{\link{plot.summary.year}} is called.}

\item{which.summary.year}{indices of the required plots in \code{\link{plot.summary.year}}.}

\item{accuracy}{numerical scalar; see \code{\link{scales::number_format}(accuracy=)}.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot,
see \code{\link{par}(ask=.)}.}
}
\value{
An S3 object of \code{\link{class}} \code{wos.db}.
A \code{list} with the folowing \code{data.frame}s:
\describe{
  \item{Docs}{}
}
Authors AutDoc Categories CatDoc Areas AreaDoc Addresses AddAutDoc Journals
}
\description{
\code{CreateDB.wos} converts a \code{data.frame} with WoS variables (see \code{\link{wosdf}}) into...
}
\examples{
db <- CreateDB.wos(wosdf)
str(db, 1)
print(db)
summary(db)
}
\seealso{
\code{\link{wosdf}}, \code{\link{ImportSources.wos}}.

\code{\link{summary.wos.db}}.

\code{\link{plot.summary.wos.db}}, \code{\link{plot.summary.year}}
}
