% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcr.R
\name{CreateDBJCR}
\alias{CreateDBJCR}
\alias{jcr.db-class}
\title{Import JCR data from Web of Science (WoS)}
\usage{
CreateDBJCR(path = ".", files = dir(path, pattern = "*.xlsx"), verbose = TRUE)
}
\arguments{
\item{path}{character; path to the directory containing the files.
Defaults to the working directory.}

\item{files}{character vector with the file names. Defaults to filenames in
directory \code{path} with extension \emph{.xlsx"}.}

\item{verbose}{logical; indicating whether the name of the file being processed
is printed. Defaults to \code{TRUE}.}
}
\value{
An S3 object of \link{class} \code{jcr.db}. A \code{list} with components:
\code{Sources}, \code{Categories}, \code{JCRSour} and \code{JCRCatSour}.
}
\description{
Reads JCR data from excel files downloaded from WoS and generates a
relational database (a list of data.frames).
It is assumed that the file name format is \emph{JCR_\link{WE}_\link{PY}}, where \emph{\link{WE}} is
the WoS index (SCIE, SSCI, ...) and \emph{\link{PY}} the JCR year.
}
\seealso{
\link{CreateDB}, \link{AddDBJCR}.
}
