% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos.R
\name{db_bib}
\alias{db_bib}
\alias{wos.db-class}
\alias{db_bib.wos.data}
\alias{print.wos.db}
\alias{plot.wos.db}
\title{Bibliographic data base creation}
\usage{
db_bib(data, ...)

\method{db_bib}{wos.data}(data, label = "", progress = interactive(), verbose = FALSE, ...)

\method{print}{wos.db}(x, ...)

\method{plot}{wos.db}(
  x,
  filter,
  which = 1:3,
  plot = TRUE,
  warning = FALSE,
  ask = plot && (length(which) > 1) && interactive(),
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} with bibliographic data.}

\item{...}{further arguments passed to or from other methods.}

\item{label}{character string describing the data.}

\item{progress}{logical; if \code{TRUE} the progress is printed.}

\item{verbose}{logical; if \code{TRUE} additional information is printed.}

\item{x}{an object used to select a method.}

\item{filter}{vector of document identifiers (usually a result of \code{\link{get_id_docs}}).}

\item{which}{if a subset of the plots is required,
specify a subset of the numbers \code{1:3}.}

\item{plot}{logical; if \code{TRUE} (default), the plots are drawn, otherwise only
the list of ggplot2 objects is (invisibly) returned.}

\item{warning}{logical; if \code{FALSE} (default), warnings are ignored.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot
(see \code{\link{par}(ask=.)}).}
}
\value{
An S3 object of \link{class} \code{wos.db}.
A \code{list} with the following components:
\itemize{
\item \code{Docs}: document-level records (year, type, citations, identifiers).
\item \code{Authors} and \code{AutDoc}: author dictionary and document-author links.
\item \code{OI} and \code{OIDoc}: ORCID identifiers and their linkage to documents.
\item \code{RI} and \code{RIDoc}: ResearcherID identifiers and their linkage to documents.
\item \code{Affiliations} and \code{AffDoc}: affiliation dictionary and document links.
\item \code{Addresses} and \code{AddAutDoc}: address and document-author-address links.
\item \code{Sources}: document sources (journals, books, proceedings, ...).
\item \code{Categories}, \code{Areas}, \code{CatSour} and \code{AreaSour}: thematic classifications
and their linkage to sources.
\item \code{WSIndex} and \code{SourWSI}: Web of Science indexes and their linkage to
sources.
\item \code{label} and \code{date}: metadata that helps tracking the dataset
identity over time.
}

This object has specialized \link{print} and \link{plot} methods.
}
\description{
\code{db_bib.wos.data()} converts a \code{data.frame} with WoS variables
(as returned by \code{\link{import_wos}}; see \code{\link{wosdf}}) into
a relational database (a list of data.frames).
}
\examples{
db <- db_bib(wosdf)
str(db, 1)
print(db)
summary(db)
}
\seealso{
\code{\link{wosdf}}, \code{\link{import_wos}}.

\code{\link{summary.wos.db}}.

\code{\link{plot.summary.wos.db}}, \code{\link{plot.summary.year.wos}}
}
