% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcr.R
\name{db_jcr}
\alias{db_jcr}
\alias{jcr.db-class}
\title{Import JCR data from Web of Science (WoS)}
\usage{
db_jcr(path = ".", files = dir(path, pattern = "*.xlsx"), verbose = TRUE)
}
\arguments{
\item{path}{character; path to the directory containing the files.
Defaults to the working directory.}

\item{files}{character vector with the file names. Defaults to filenames in
directory \code{path} with extension \emph{.xlsx"}.}

\item{verbose}{logical; indicating whether the name of the file being processed
is printed. Defaults to \code{TRUE}.}
}
\value{
An S3 object of \link{class} \code{jcr.db}. A \code{list} with components:
\code{Sources}, \code{Categories}, \code{JCRSour} and \code{JCRCatSour}.
}
\description{
Reads JCR data from excel files downloaded from WoS and generates a
relational database (a list of data.frames).
It is assumed that the file name format is \emph{JCR_[WE]_[PY]}, where \emph{[WE]} is
the WoS index (SCIE, SSCI, ...) and \emph{[PY]} the JCR year.
}
\details{
A subscription to Web of Science is required to download JCR data.
For successful execution, Excel files must retain the original header fields
from Clarivate exports:
\code{Title20}, \code{ISO_ABBREV}, \code{TITLE}, \code{ISSN}, \code{EISSN}, \code{ISSUES/YEAR}, \code{COUNTRY},
\code{LANGUAGE}, \verb{1ST_YR_PUB}, \code{categories}, \code{TOT_CITES}, \code{CITES_JCR_YR}, \code{CITES_JCR_YR1},
\code{CITES_JCR_YR2}, \code{IF_NUMERATOR}, \code{CITES_JCR_YR3}, \code{CITES_JCR_YR4}, \code{CITES_JCR_YR5},
\verb{5YR_IF_NUMERATOR}, \code{ITEMS_JCR_YR}, \code{ITEMS_JCR_YR1}, \code{ITEMS_JCR_YR2}, \code{IF_DENOMINATOR},
\code{ITEMS_JCR_YR3}, \code{ITEMS_JCR_YR4}, \code{ITEMS_JCR_YR5}, \verb{5YR_IF_DENOMINATOR}, \code{IMPACT_FACTOR},
\code{IMMEDIACY_INDEX}, \code{CITED_HALF_LIFE}, \verb{5YR_IMPACT_FACTOR}, \code{EIGENFACTOR},
\code{NORM_EIGENFACTOR}, \code{ARTL_INFLUENCE}, \code{FREQUENCY}, \code{PUBCODE}, \code{PUBLISHER_NAME},
\code{CATEGORY_CODE}, \code{CATEGORY_DESCRIPTION}, \code{CATEGORY_RANKING}, \code{QUARTILE_RANK}, \code{JIF_PERCENTILE}.
}
\seealso{
\link{db_bib}, \link{add_jcr}.
}
