% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimetr_package.R
\docType{data}
\name{dbjcr}
\alias{dbjcr}
\title{Mathematics UDC 2018-2023 bibliographic database}
\format{
An object of class \code{wos.jcr} (inherits from \code{wos.db}) of length 22.
}
\usage{
dbjcr
}
\description{
Bibliographic database with JCR metrics (a \linkS4class{wos.jcr} S3 object)
corresponding to a WoS search by the Affiliation field of \emph{Universidade da Coruña (UDC)}
in the research area \code{"Mathematics"} during the years 2018–2023
(generated fom \link{wosdf} data set, using the functions \link{db_bib}, \link{db_jcr}
and \link{add_jcr}).
}
\keyword{datasets}
