% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getid.R
\name{get_id_docs}
\alias{get_id_docs}
\alias{get_id_authors}
\alias{get_id_addresses}
\alias{get_id_areas}
\alias{get_id_categories}
\alias{get_id_wsi}
\alias{get_id_sources}
\alias{get_id_oi}
\alias{get_id_ri}
\alias{get_id_affiliations}
\title{Get table identifiers}
\usage{
get_id_docs(
  db,
  ...,
  id_sources,
  id_authors,
  id_addresses,
  id_oi,
  id_ri,
  id_affiliations
)

get_id_authors(db, ...)

get_id_addresses(db, ...)

get_id_areas(db, ...)

get_id_categories(db, ...)

get_id_wsi(db, ...)

get_id_sources(db, ..., id_categories, id_areas, id_wsi)

get_id_oi(db, ...)

get_id_ri(db, ...)

get_id_affiliations(db, ...)
}
\arguments{
\item{db}{Object of \code{\link{class}} \code{wos.db}, as returned by \link{db_bib}.}

\item{...}{Logical predicates. Multiple conditions are combined with \code{&}
(see \code{\link[dplyr]{filter}}).}

\item{id_sources}{optional; sources identifiers.}

\item{id_authors}{optional; author identifiers (values of \code{db$Authors$ids}).}

\item{id_addresses}{optional; addresses identifiers.}

\item{id_oi}{optional; ORCID identifiers or codes (values of \code{db$OI$OI}).}

\item{id_ri}{optional; RI identifiers or codes (values of \code{db$RI$RI}).}

\item{id_affiliations}{optional; affiliations identifiers or names (values of \code{db$Affiliations$C3}).}

\item{id_categories}{optional; categories identifiers or names (values of \code{db$Categories$WC})}

\item{id_areas}{optional; research area identifiers or names (values of \code{db$Areas$SC})}

\item{id_wsi}{optional; WoS Index identifiers or names (values of \code{db$WSI$WE})}
}
\value{
An integer vector of identifiers.
For instance, \code{get_id_docs} returns the document identifiers (values of
the \code{db$Docs$idd} variable) corresponding to identifiers of authors,
categories, areas, addresses and/or sources.
Logical expressions defined in terms of the variables in \code{db$Docs} can
also be used as arguments.
}
\description{
Retrieve internal identifiers (entity keys) for the corresponding relational
table (authors, sources, categories, areas, ...). Any field in the target
table can be used as a condition, and multiple conditions are combined with \code{&}
(only IDs for which all conditions evaluate to TRUE are returned).
\emph{Tidy evaluation} can be used to construct the logical expressions,
see \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\seealso{
\link{db_bib}, \code{\link[dplyr]{filter}}.
}
