% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_jcr.R
\name{get_jcr}
\alias{get_jcr}
\alias{get_jcr_cat}
\title{Get JCR metrics}
\usage{
get_jcr(db, filter)

get_jcr_cat(db, filter, best = TRUE)
}
\arguments{
\item{db}{a bibliographic database with JCR information (a \linkS4class{wos.jcr} object;
typically the output of the function \link{AddDBJCR}).}

\item{filter}{vector of document identifiers (usually a result of \link{get.idDocs}).}

\item{best}{logical; if \code{TRUE} (default), only the results for the WoS category
with the best ranking for each document are returned.}
}
\description{
\code{get_jcr} combines document indexes with their source JCR metrics per year.

\code{get_jcr_cat} combines document indexes with their source JCR metrics per year
and WoS category.
}
\seealso{
\link{AddDBJCR}, \link{CreateDBJCR}, \link{CreateDB}.
}
