% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wosdb.R
\name{ImportDB.wos}
\alias{ImportDB.wos}
\title{Import DB}
\usage{
ImportDB.wos(dbname, label = dbname, drv = RSQLite::SQLite(), ...)
}
\arguments{
\item{dbname}{The path to the database file.}

\item{label}{character string describing the data.}

\item{drv}{DBMS driver (e.g., `RSQLite()`, `MySQL()`, `PostgreSQL()`...)}

\item{...}{authentication arguments needed by the DBMS instance;
these typically include `user`, `password`, `host`, `port`, etc.
For details see the appropriate DBIDriver.}
}
\value{
An S3 object of \code{\link{class}} \code{wos.db}.
A \code{list} with the folowing \code{data.frame}s:
\describe{
  \item{Docs}{}
}
Authors AutDoc Categories CatDoc Areas AreaDoc Addresses AddAutDoc Sources
}
\description{
`ImportDB.wos` imports tables from remote DataBase Management System (DBMS)...
}
\examples{
db <- CreateDB.wos(wosdf)
str(db, 1)
print(db)
summary(db)
}
\seealso{
\code{\link{wosdf}}, , \code{\link{ImportSources.wos}}, \code{\link{CreateDB.wos}}.
}
