% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos.R
\name{ImportSources.wos}
\alias{ImportSources.wos}
\title{Import bibliographic data downloaded from Web of Science (WoS).}
\usage{
ImportSources.wos(path = ".", pattern = "*.txt", all = TRUE, progress = NULL)
}
\arguments{
\item{path}{character; path to the directory containing the files.}

\item{pattern}{regular expression; only matching files will be loaded.
Defaults to \code{"*.txt"}.}

\item{all}{logical; indicating whether sources without ISSN are included
(\code{TRUE}) or not (\code{FALSE}) in the result. Defaults to \code{TRUE}.}

\item{progress}{logical; indicating whether a progress bar should be displayed.
Defaults to \code{TRUE} when R is used interactively and there is more than one file,
and \code{FALSE} otherwise.}
}
\value{
A \code{data.frame} with rows corresponding to sources and columns to
WoS variables.
}
\description{
Reads bibliography entries from UTF-8 encoded Tab-delimited files containing
"Full Record and Cited References" (see \link{wosdf}).
}
\seealso{
\link{wosdf}, \link{CreateDB}.
}
