% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adist_authors.R
\name{match_authors}
\alias{match_authors}
\title{Approximate Authors' Names Matching}
\usage{
match_authors(authors, table, weight = c(family = 0.9, name = 0.1),
  max.dist = 0.2, attr.dist = FALSE)
}
\arguments{
\item{authors}{author names to be approximately matched.}

\item{table}{lookup table for matching.}

\item{weight}{weights associated with differences in family names and in first names.}

\item{max.dist}{maximum distance between strings to be matched.}

\item{attr.dist}{logical; indicating whether to optionally return the distances
(weighted proporpion of names coincidences) as the \code{"dist"} attribute of the return value.}
}
\value{
Returns a vector of the positions of the closest matches of its first argument in its second.
When multiple matches with the same smallest distance exist, the first one is returned.
When no match is found \code{NA_integer_} is returned.
}
\description{
Approximate string matching between character vectors containing authors' names
(equivalents of R's native \code{\link{match}}).
}
\seealso{
\code{\link{adist_authors}}.
}
