% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimetr_internals.R
\name{scimetr-internals}
\alias{scimetr-internals}
\alias{.stable}
\title{scimetr internal and secondary functions}
\usage{
.stable(x, nmax = 10, others = TRUE)
}
\description{
Listed below are supporting functions for the major methods in scimetr.
}
\keyword{internal}
