% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.wos.db}
\alias{summary.wos.db}
\alias{print.summary.wos.db}
\alias{TC.authors}
\alias{plot.summary.wos.db}
\alias{summary.wos.jcr}
\alias{print.summary.wos.jcr}
\alias{plot.summary.wos.jcr}
\title{Summaries of a bibliographic database}
\usage{
\method{summary}{wos.db}(object, filter, TC.index = c("H", "G"), nmax = 10, ...)

\method{print}{summary.wos.db}(x, ...)

TC.authors(db, idAuthors)

\method{plot}{summary.wos.db}(
  x,
  which = 1:5,
  pie = FALSE,
  others = !pie,
  accuracy = 0.1,
  plot = TRUE,
  ask = length(which) > 1 && interactive(),
  ...
)

\method{summary}{wos.jcr}(object, filter, all = FALSE, ...)

\method{print}{summary.wos.jcr}(x, digits = 2, all = TRUE, ...)

\method{plot}{summary.wos.jcr}(x, plot = TRUE, all = FALSE, ask = plot && all && interactive(), ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{filter}{vector of document identifiers
(optional, usually a result of \link{get.idDocs}).}

\item{TC.index}{citation indexes.}

\item{nmax}{number of top levels.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object used to select a method.}

\item{db}{Object of \code{\link{class}} \code{wos.db}, as returned by \link{CreateDB}.}

\item{idAuthors}{optional; author identifiers}

\item{which}{if a subset of the plots is required,
specify a subset of the numbers \code{1:5}.}

\item{pie}{logical; if \code{TRUE}, pie charts are drawn.}

\item{others}{logical; if \code{FALSE}, only \code{nmax} top levels categories are shown.}

\item{accuracy}{numerical scalar; see \code{\link[scales]{label_number}(accuracy=.)}.}

\item{plot}{logical; if \code{TRUE} (default), the plots are drawn, otherwise only
the list of ggplot2 objects is (invisibly) returned.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot,
see \code{\link{par}(ask=.)}.}

\item{all}{logical; if \code{TRUE}, the corresponding \linkS4class{wos.db} method
is called (additional parameters \code{...} are passed to it), otherwise only results
of JCR metrics are returned.}

\item{digits}{minimal number of significant digits, see \link{print.default}.}
}
\description{
\link{summary} methods for WoS bibliometric databases.
}
\seealso{
\link{CreateDB}, \link{AddDBJCR}.
}
