% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_year.R
\name{summary_year}
\alias{summary_year}
\alias{summary_year.wos.db}
\alias{print.summary.year.wos}
\alias{plot.summary.year.wos}
\alias{summary_year.wos.jcr}
\alias{print.summary.year.jcr}
\alias{plot.summary.year.jcr}
\title{Summaries per year}
\usage{
summary_year(object, ...)

\method{summary_year}{wos.db}(object, filter, ...)

\method{print}{summary.year.wos}(x, ...)

\method{plot}{summary.year.wos}(
  x,
  which = 1:3,
  boxplot = FALSE,
  plot = TRUE,
  ask = plot && length(which) > 1 && interactive(),
  ...
)

\method{summary_year}{wos.jcr}(object, filter, all = FALSE, ...)

\method{print}{summary.year.jcr}(x, digits = 3, all = TRUE, ...)

\method{plot}{summary.year.jcr}(x, plot = TRUE, all = FALSE, ask = plot && interactive(), ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{further arguments passed to or from other methods.}

\item{filter}{vector of document identifiers
(optional, usually a result of \link{get.idDocs}).}

\item{x}{an object used to select a method.}

\item{which}{if a subset of the plots is required,
specify a subset of the numbers \code{1:3}.}

\item{boxplot}{logical; if \code{TRUE}, boxplots are drawn (for plots from 2 to 3).}

\item{plot}{logical; if \code{TRUE} (default), the plots are drawn, otherwise only
the list of ggplot2 objects is (invisibly) returned.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot,
see \code{\link{par}(ask=.)}.}

\item{all}{logical; if \code{TRUE}, the corresponding \linkS4class{wos.db} method
is called (additional parameters \code{...} are passed to it), otherwise only results
of JCR metrics are returned.}
}
\description{
The \code{summary_year} methods return summaries by year of a bibliometric database.
}
\seealso{
\link{CreateDB}, \link{AddDBJCR}.
}
