% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimetr_package.R
\docType{data}
\name{wosdf}
\alias{wosdf}
\title{Mathematics UDC WoS Core Collection 2018-2023 data}
\format{
A data frame with 293 rows and 48 columns:
\describe{
\item{PT}{Publication Type}
\item{AU}{Author}
\item{AF}{Author Full Name}
\item{TI}{Article Title}
\item{SO}{Source Title}
\item{SE}{Book Series Title}
\item{BS}{Book Series Subtitle}
\item{LA}{Language}
\item{DT}{Document Type}
\item{C1}{Address}
\item{C3}{Affiliation}
\item{RI}{Researcher Ids}
\item{OI}{ORCID}
\item{NR}{Cited Reference Count}
\item{TC}{Times Cited, WoS Core}
\item{Z9}{Times Cited, All Databases}
\item{U1}{180 Day Usage Count}
\item{U2}{Since 2013 Usage Count}
\item{PU}{Publisher}
\item{PI}{Publisher City}
\item{PA}{Publisher Address}
\item{SN}{ISSN}
\item{EI}{eISSN}
\item{BN}{ISBN}
\item{J9}{Journal Abbreviation}
\item{JI}{Journal ISO Abbreviation}
\item{PD}{Publication Date}
\item{PY}{Publication Year}
\item{VL}{Volume}
\item{IS}{Issue}
\item{PN}{Part Number}
\item{SU}{Supplement}
\item{SI}{Special Issue}
\item{MA}{Meeting Abstract}
\item{BP}{Start Page}
\item{EP}{End Page}
\item{AR}{Article Number}
\item{DI}{DOI}
\item{D2}{Book DOI}
\item{EA}{Early Access Date}
\item{PG}{Number of Pages}
\item{WC}{WoS Category}
\item{WE}{Web of Science Index}
\item{SC}{Research Area}
\item{HC}{Highly Cited Status}
\item{HP}{Hot Paper Status}
\item{DA}{Date of Export}
\item{UT}{Unique WOS ID}
}
}
\source{
Clarivate Web of Science:
\url{https://www.webofscience.com/}.
}
\usage{
wosdf
}
\description{
The data set consists of 293 publications corresponding to a WoS search by
the Affiliation field of \emph{Universidade da Coruña (UDC)} (Affiliation:
OG = Universidade da Coruna) in the research area \code{"Mathematics"} (\code{SC = Mathematics}),
indexed in the Web of Science Core Collection:
\itemize{
\item{Science Citation Index Expanded (SCI-EXPANDED).}
\item{Social Sciences Citation Index (SSCI).}
\item{Arts & Humanities Citation Index (A&HCI).}
}
in the years 2018-2023 (generated using the \link{import_wos} function).
}
\seealso{
\link{import_wos}, \link{db_bib}
}
\keyword{datasets}
