% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimetr_package.R
\docType{data}
\name{wosdf2}
\alias{wosdf2}
\title{Mathematics UDC WoS Core Collection 2008-2017 data}
\format{A data frame with 389 observations on the following 62 variables:
\describe{
  \item{PT}{Publication type.}
  \item{AU}{Author.}
  \item{BA}{Book authors.}
  \item{BE}{Editor.}
  \item{GP}{Group author.}
  \item{AF}{Author full.}
  \item{BF}{Book authors fullname.}
  \item{CA}{Corporate author.}
  \item{TI}{Title.}
  \item{SO}{Publication name.}
  \item{SE}{Series title.}
  \item{BS}{Book series.}
  \item{LA}{Language.}
  \item{DT}{Document type.}
  \item{CT}{Conference title.}
  \item{CY}{Conference year.}
  \item{CL}{Conference place.}
  \item{SP}{Conference sponsors.}
  \item{HO}{Conference host.}
  \item{DE}{Keywords.}
  \item{ID}{Keywords Plus.}
  \item{AB}{Abstract.}
  \item{C1}{Addresses.}
  \item{RP}{Reprint author.}
  \item{EM}{Author email.}
  \item{RI}{Researcher id numbers.}
  \item{OI}{Orcid numbers.}
  \item{FU}{Funding agency and grant number.}
  \item{FX}{Funding text.}
  \item{CR}{Cited references.}
  \item{NR}{Number of cited references.}
  \item{TC}{Times cited.}
  \item{Z9}{Total times cited count .}
  \item{U1}{Usage Count (Last 180 Days).}
  \item{U2}{Usage Count (Since 2013).}
  \item{PU}{Publisher.}
  \item{PI}{Publisher city.}
  \item{PA}{Publisher address.}
  \item{SN}{ISSN.}
  \item{EI}{eISSN.}
  \item{BN}{ISBN.}
  \item{J9}{Journal.ISI.}
  \item{JI}{Journal.ISO.}
  \item{PD}{Publication date.}
  \item{PY}{Year published.}
  \item{VL}{Volume.}
  \item{IS}{Issue.}
  \item{PN}{Part number.}
  \item{SU}{Supplement.}
  \item{SI}{Special issue.}
  \item{MA}{Meeting abstract.}
  \item{BP}{Beginning page.}
  \item{EP}{Ending page.}
  \item{AR}{Article number.}
  \item{DI}{DOI.}
  \item{D2}{Book DOI.}
  \item{PG}{Page count.}
  \item{WC}{WOS category.}
  \item{SC}{Research areas.}
  \item{GA}{Document delivery number.}
  \item{UT}{Access number.}
  \item{PM}{Pub Med ID.}
}}
\source{
Thomson Reuters' Web of Science: \cr
\url{http://www.webofknowledge.com}.
}
\description{
The data set consists of 389 registros vinculados a la UDC por el campo Organización-Nombre preferido
(Organization-Enhaced: OG = Universidade da Coruna)
del área de investigación 'Mathematics' (Research areas: SC = \*Mathematics\*)
de los siguientes índices de la Web of Science Core Collection:
\itemize{
  \item{Science Citation Index Expanded (SCI-EXPANDED).}
  \item{Social Sciences Citation Index (SSCI).}
  \item{Arts & Humanities Citation Index (A&HCI).}
}
correspondientes a los años 2008-2017.
}
\keyword{datasets}
