% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getid.R
\name{get.idDocs}
\alias{get.idDocs}
\alias{get.idAuthors}
\alias{get.idAddresses}
\alias{get.idAreas}
\alias{get.idCategories}
\alias{get.idJournals}
\title{Get table identifiers}
\usage{
get.idDocs(db, idAuthors, idCategories, idAreas, idAddresses, idJournals)

get.idAuthors(db, ...)

get.idAddresses(db, ...)

get.idAreas(db, ...)

get.idCategories(db, ...)

get.idJournals(db, ...)
}
\arguments{
\item{db}{Object of \code{\link{class}} \code{wos.db}, as returned by \code{\link{wos_CreateDB}}.}

\item{idAuthors}{optional; author identifiers}

\item{idCategories}{optional; categories identifiers}

\item{idAreas}{optional; areas identifiers}

\item{idAddresses}{optional; addresses identifiers}

\item{idJournals}{optional; journals identifiers}

\item{...}{Logical predicates. Multiple conditions are combined with & (see
\code{\link[dplyr]{filter}})}
}
\description{
\code{get.idDocs} returns the document identifiers (values of the \code{db$Docs$idd} variable)
corresponding to identifiers of authors (values of \code{db$Authors$ids}), categories (...), areas (...),
addresses (...) and/or journals (...). Multiple conditions are combined with &.
}
\seealso{
\code{\link{wos_CreateDB}}, \code{\link[dplyr]{filter}}.
}
