% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getid.R
\name{get.idDocs}
\alias{get.idDocs}
\alias{get.idAuthors}
\alias{get.idAddresses}
\alias{get.idAreas}
\alias{get.idCategories}
\alias{get.idSources}
\title{Get table identifiers}
\usage{
get.idDocs(db, idAuthors, idCategories, idAreas, idAddresses, idSources)

get.idAuthors(db, ...)

get.idAddresses(db, ...)

get.idAreas(db, ...)

get.idCategories(db, ...)

get.idSources(db, ...)
}
\arguments{
\item{db}{Object of \code{\link{class}} \code{wos.db}, as returned by \code{\link{CreateDB.wos}}.}

\item{idAuthors}{optional; author identifiers}

\item{idCategories}{optional; categories identifiers}

\item{idAreas}{optional; areas identifiers}

\item{idAddresses}{optional; addresses identifiers}

\item{idSources}{optional; sources identifiers}

\item{...}{Logical predicates. Multiple conditions are combined with & (see
\code{\link[dplyr]{filter}})}
}
\description{
\code{get.idDocs} returns the document identifiers (values of the \code{db$Docs$idd} variable)
corresponding to identifiers of authors (values of \code{db$Authors$ids}), categories (...), areas (...),
addresses (...) and/or sources (...). Multiple conditions are combined with &.
}
\seealso{
\code{\link{CreateDB.wos}}, \code{\link[dplyr]{filter}}.
}
