% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_jcr.R
\name{get_jcr}
\alias{get_jcr}
\alias{get_jcr_cat}
\title{Get JCR metrics}
\usage{
get_jcr(db, filter)

get_jcr_cat(db, filter, best = TRUE)
}
\arguments{
\item{db}{a bibliographic database with JCR information (a \linkS4class{wos.jcr} object;
typically the output of the function \code{\link[=add_jcr]{add_jcr()}}).}

\item{filter}{vector of document identifiers (usually a result of \code{\link[=get_id_docs]{get_id_docs()}}).}

\item{best}{logical; if \code{TRUE} (default), only the results for the WoS category
with the best ranking for each document are returned.}
}
\value{
A \code{data.frame} with document indexes and their source JCR metrics.
}
\description{
\code{get_jcr()} combines document indexes with their source JCR metrics per year.

\code{get_jcr_cat()} combines document indexes with their source JCR metrics per year
and WoS category.
}
\examples{
head(get_jcr(dbjcr))
head(get_jcr_cat(dbjcr, best = TRUE))
}
\seealso{
\code{\link[=add_jcr]{add_jcr()}}, \code{\link[=db_jcr]{db_jcr()}}, \code{\link[=db_bib]{db_bib()}}.
}
