% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos.R
\name{ImportSources.wos}
\alias{ImportSources.wos}
\title{Import bibliographic data downloaded from Web of Science (WoS).}
\usage{
ImportSources.wos(
  path = ".",
  pattern = "*.txt",
  other = c("include", "ignore", "split")
)
}
\arguments{
\item{path}{to a subdir...}

\item{pattern}{file pattern (= '*.txt' by default)}

\item{other}{Guardar otros documentos "include", "ignore", "split", pendiente: cambiar split por path}
}
\value{
A \code{data.frame} with columns corresponding to WoS variables.
}
\description{
Reads bibliography entries from UTF-8 encoded Tab-delimited files containing "Full Record and Cited References"
(see \code{\link{wosdf}}.
}
\seealso{
\code{\link{wosdf}}, \code{\link{CreateDB.wos}}.
}
