% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getid.R
\name{scimetr-internals}
\alias{scimetr-internals}
\alias{.get.idDocs2}
\title{scimetr internal and secondary functions}
\usage{
.get.idDocs2(
  db,
  idAuthors,
  idCategories,
  idAreas,
  idAddresses,
  idSources,
  indices = TRUE
)
}
\description{
Listed below are supporting functions for the major methods in scimetr.
}
\keyword{internal}
