% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimetr_package.R
\docType{package}
\name{scimetr-package}
\alias{scimetr-package}
\alias{scimetr}
\title{scimetr: Analysis of Scientific Publication Data with R}
\description{
This package implements tools for quantitative research in scientometrics and bibliometrics.
It provides routines for importing bibliographic data from
Thomson Reuters' Web of Knowledge (http://www.webofknowledge.com) and performing bibliometric analysis.
For more information visit https://rubenfcasal.github.io/scimetr/articles/scimetr.html.
}
