% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.wos.db}
\alias{summary.wos.db}
\alias{print.summary.wos.db}
\alias{TC.authors}
\alias{plot.summary.wos.db}
\title{WoS DB Summaries}
\usage{
\method{summary}{wos.db}(object, filter, TC.index = c("H", "G"),
  nmax = 10, ...)

\method{print}{summary.wos.db}(x, ...)

TC.authors(db, idAuthors)

\method{plot}{summary.wos.db}(x, which = 1:5, pie = FALSE,
  others = !pie, accuracy = 0.1, ask = length(which) > 1 &&
  dev.interactive(), ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{filter}{vector of document identifiers (usually a result of \code{\link{get.idDocs}}).}

\item{TC.index}{citation indexes.}

\item{nmax}{number of top levels.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object used to select a method.}

\item{db}{Object of \code{\link{class}} \code{wos.db}, as returned by \code{\link{wos_CreateDB}}.}

\item{idAuthors}{optional; author identifiers}

\item{which}{if a subset of the plots is required,
specify a subset of the numbers \code{1:5}.}

\item{pie}{logical; if \code{TRUE}, pie charts are drawn.}

\item{others}{logical; if \code{FALSE}, only \code{nmax} top levels categories are shown.}

\item{accuracy}{numerical scalar; see \code{\link{scales::number_format}(accuracy=)}.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot,
see \code{\link{par}(ask=.)}.}
}
\description{
\code{summary} methods for \code{wos.db}-class...
}
\seealso{
\code{\link{wos_CreateDB}}.
}
