% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.year.R
\name{summary_year}
\alias{summary_year}
\alias{print.summary.year}
\alias{plot.summary.year}
\title{Summary year}
\usage{
summary_year(object, filter, ...)

\method{print}{summary.year}(x, ...)

\method{plot}{summary.year}(
  x,
  which = 1:3,
  boxplot = FALSE,
  ask = length(which) > 1 && dev.interactive(),
  ...
)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{filter}{vector of document identifiers (usually a result of \code{\link{get.idDocs}}).}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object used to select a method.}

\item{which}{if a subset of the plots is required,
specify a subset of the numbers \code{1:3}.}

\item{boxplot}{logical; if \code{TRUE}, boxplots are drawn (for plots from 2 to 4).}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot,
see \code{\link{par}(ask=.)}.}
}
\description{
\code{summary_year} returns a summary per year...
}
\seealso{
\code{\link{CreateDB.wos}}.
}
